# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DLJAIS AI Action OS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Conversation {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
  /** @nullable */
  lastMessage?: string | null;
}

export interface ConversationInput {
  title: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  /** @nullable */
  intent?: string | null;
  createdAt: string;
  /** @nullable */
  actionCardId?: number | null;
}

export interface ConversationDetail {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messages: Message[];
}

export interface MessageInput {
  content: string;
}

export type ActionCardStatus = typeof ActionCardStatus[keyof typeof ActionCardStatus];


export const ActionCardStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  executing: 'executing',
  completed: 'completed',
  failed: 'failed',
} as const;

export type ActionCardRiskLevel = typeof ActionCardRiskLevel[keyof typeof ActionCardRiskLevel];


export const ActionCardRiskLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface ActionCard {
  id: number;
  title: string;
  platform: string;
  intent: string;
  status: ActionCardStatus;
  riskLevel: ActionCardRiskLevel;
  /** @nullable */
  estimatedCost?: string | null;
  details?: string;
  /** @nullable */
  preview?: string | null;
  /** @nullable */
  conversationId?: number | null;
  createdAt: string;
  updatedAt?: string;
}

export interface MessagePair {
  userMessage: Message;
  aiMessage: Message;
  actionCard?: ActionCard;
}

export type ActionInputRiskLevel = typeof ActionInputRiskLevel[keyof typeof ActionInputRiskLevel];


export const ActionInputRiskLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface ActionInput {
  title: string;
  platform: string;
  intent: string;
  riskLevel: ActionInputRiskLevel;
  estimatedCost?: string;
  details: string;
  preview?: string;
  conversationId?: number;
}

export type ActionStatusUpdateStatus = typeof ActionStatusUpdateStatus[keyof typeof ActionStatusUpdateStatus];


export const ActionStatusUpdateStatus = {
  approved: 'approved',
  rejected: 'rejected',
  executing: 'executing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface ActionStatusUpdate {
  status: ActionStatusUpdateStatus;
  modifiedDetails?: string;
}

export type PlatformCategory = typeof PlatformCategory[keyof typeof PlatformCategory];


export const PlatformCategory = {
  social: 'social',
  ads: 'ads',
  ecommerce: 'ecommerce',
  trading: 'trading',
  food: 'food',
  website: 'website',
} as const;

export type PlatformStatus = typeof PlatformStatus[keyof typeof PlatformStatus];


export const PlatformStatus = {
  connected: 'connected',
  disconnected: 'disconnected',
  limited: 'limited',
} as const;

export interface Platform {
  id: number;
  name: string;
  category: PlatformCategory;
  status: PlatformStatus;
  icon: string;
  /** @nullable */
  accountName?: string | null;
  /** @nullable */
  lastSync?: string | null;
}

export type PlatformUpdateStatus = typeof PlatformUpdateStatus[keyof typeof PlatformUpdateStatus];


export const PlatformUpdateStatus = {
  connected: 'connected',
  disconnected: 'disconnected',
  limited: 'limited',
} as const;

export interface PlatformUpdate {
  status: PlatformUpdateStatus;
  accountName?: string;
}

export interface StatsSummary {
  totalActions: number;
  pendingActions: number;
  completedActions: number;
  connectedPlatforms: number;
  actionsThisWeek: number;
  recentActivity?: ActionCard[];
}

export interface PendingCount {
  count: number;
}

export type ListActionsParams = {
status?: ListActionsStatus;
};

export type ListActionsStatus = typeof ListActionsStatus[keyof typeof ListActionsStatus];


export const ListActionsStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  executing: 'executing',
  completed: 'completed',
  failed: 'failed',
} as const;