# lib/api-zod/src/generated/types/actionCard.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DLJAIS AI Action OS API
 * OpenAPI spec version: 0.1.0
 */
import type { ActionCardRiskLevel } from './actionCardRiskLevel';
import type { ActionCardStatus } from './actionCardStatus';

export interface ActionCard {
  id: number;
  title: string;
  platform: string;
  intent: string;
  status: ActionCardStatus;
  riskLevel: ActionCardRiskLevel;
  /** @nullable */
  estimatedCost?: string | null;
  details?: string;
  /** @nullable */
  preview?: string | null;
  /** @nullable */
  conversationId?: number | null;
  createdAt: string;
  updatedAt?: string;
}