# lib/api-zod/src/generated/types/actionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DLJAIS AI Action OS API
 * OpenAPI spec version: 0.1.0
 */
import type { ActionInputRiskLevel } from './actionInputRiskLevel';

export interface ActionInput {
  title: string;
  platform: string;
  intent: string;
  riskLevel: ActionInputRiskLevel;
  estimatedCost?: string;
  details: string;
  preview?: string;
  conversationId?: number;
}