# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DLJAIS AI Action OS API
 * OpenAPI spec version: 0.1.0
 */
import type { MessageRole } from './messageRole';

export interface Message {
  id: number;
  conversationId: number;
  role: MessageRole;
  content: string;
  /** @nullable */
  intent?: string | null;
  createdAt: string;
  /** @nullable */
  actionCardId?: number | null;
}