# lib/api-zod/src/generated/types/platform.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DLJAIS AI Action OS API
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformCategory } from './platformCategory';
import type { PlatformStatus } from './platformStatus';

export interface Platform {
  id: number;
  name: string;
  category: PlatformCategory;
  status: PlatformStatus;
  icon: string;
  /** @nullable */
  accountName?: string | null;
  /** @nullable */
  lastSync?: string | null;
}