# lib/api-zod/src/generated/types/platformUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DLJAIS AI Action OS API
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformUpdateStatus } from './platformUpdateStatus';

export interface PlatformUpdate {
  status: PlatformUpdateStatus;
  accountName?: string;
}