# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DLJAIS AI Action OS API
 * OpenAPI spec version: 0.1.0
 */
import type { ActionCard } from './actionCard';

export interface StatsSummary {
  totalActions: number;
  pendingActions: number;
  completedActions: number;
  connectedPlatforms: number;
  actionsThisWeek: number;
  recentActivity?: ActionCard[];
}